// src/interface.ts
export interface ILiveRoom {
  id?: number;
  /** 标题 */
  title?: string;
  /** 简介 */
  desc?: string;
  /** 权重 */
  priority?: number;
  /** 推流秘钥 */
  key?: string;
  /** 类型 */
  type?: LiveRoomTypeEnum;
  /** 封面图 */
  cover_img?: string;
  /** 关键帧图 */
  keyframe_img?: string;
  /** 背景图 */
  bg_img?: string;
  /** 状态 */
  status?: LiveRoomStatusEnum;
  /** 是否显示 */
  is_show?: SwitchEnum;
  /** 是否开启聊天 */
  open_chat?: SwitchEnum;
  /** 提醒游客登录 */
  tip_tourist_login?: LiveRoomTipTouristLoginEnum;
  /** 提醒游客登录间隔 */
  tip_tourist_login_delay?: number;
  /** 游客能否发言 */
  tourist_send_msg?: SwitchEnum;
  /** 聊天关键词过滤 */
  keyword_filter_msg?: string;
  /** 用户名关键词过滤 */
  keyword_filter_username?: string;
  /** 历史消息条数 */
  history_msg_total?: number;
  /** 新注册用户不能发言时间(秒) */
  newuser_send_msg_delay?: number;
  /** 房间密码 */
  room_pwd?: string;
  /** 关闭房间 */
  is_close?: SwitchEnum;
  /** 关闭房间描述 */
  is_close_desc?: string;
  /** 发送消息间隔(秒) */
  send_msg_throttle?: number;
  /** 是否开启签到 */
  is_show_signin?: SwitchEnum;
  /** 是否开启手机看直播 */
  is_show_phone_live?: SwitchEnum;
  /** 公告通知 */
  notice_msg?: string;
  /** 系统消息 */
  system_msg?: string;
  /** 显示在线人数 */
  is_show_live_user_nums?: SwitchEnum;
  /** 设置最低在线人数 */
  mock_live_user_nums_min?: number;
  /** 设置最高在线人数 */
  mock_live_user_nums_max?: number;
  /** 最在线人数刷新间隔 */
  mock_live_user_nums_refresh_delay?: number;
  /** 聊天消息审核 */
  msg_verify?: SwitchEnum;
  is_fake?: SwitchEnum;

  pull_rtmp_url?: string;
  pull_flv_url?: string;
  pull_hls_url?: string;
  pull_webrtc_url?: string;

  pull_cdn_rtmp_url?: string;
  pull_cdn_flv_url?: string;
  pull_cdn_hls_url?: string;
  pull_cdn_webrtc_url?: string;

  push_rtmp_url?: string;
  push_obs_server?: string;
  push_obs_stream_key?: string;
  push_webrtc_url?: string;
  push_srt_url?: string;

  push_cdn_rtmp_url?: string;
  push_cdn_obs_server?: string;
  push_cdn_obs_stream_key?: string;
  push_cdn_webrtc_url?: string;
  push_cdn_srt_url?: string;

  forward_bilibili_url?: string;
  forward_douyu_url?: string;
  forward_huya_url?: string;
  forward_douyin_url?: string;
  forward_kuaishou_url?: string;
  forward_xiaohongshu_url?: string;

  /** 直播间备注 */
  remark?: string;

  areaId?: number;

  /** 用户信息 */
  user?: IUser;
  /** 用户信息 */
  users?: IUser[];
  /** 分区信息 */
  area?: IArea;
  /** 分区信息 */
  areas?: IArea[];
  /** 直播信息 */
  live?: ILive;
  user_live_room?: IUserLiveRoom & { user: IUser };

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IRedbagSend {
  id?: number;

  user_id?: number;
  live_room_id?: number;

  total_amount?: string;
  remaining_amount?: string;
  total_nums?: number;
  remaining_nums?: number;
  remark?: string;

  /** 用户信息 */
  user?: IUser;
  /** 直播间信息 */
  live_room?: IGoods;

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export enum DanmuMsgTypeEnum {
  danmu,
  otherJoin,
  userLeaved,
  system,
  redbag,
  reward,
}

export enum MsgIsFileEnum {
  yes,
  no,
}

export enum MsgContentTypeEnum {
  txt,
  img,
  video,
}

export enum MsgIsShowEnum {
  yes,
  no,
}

export enum MsgIsVerifyEnum {
  yes,
  no,
}

export interface IMsg {
  id?: number;
  live_record_id?: number;
  username?: string;
  origin_username?: string;
  content_type?: MsgContentTypeEnum;
  content?: string;
  origin_content?: string;
  live_room_id?: number;
  user_id?: number;
  client_ip?: string;
  client_env?: ClientEnvEnum;
  client_app?: ClientAppEnum;
  client_app_version?: string;
  msg_type?: DanmuMsgTypeEnum;
  user_agent?: string;
  send_msg_time?: number;
  is_show?: SwitchEnum;
  is_bilibili?: SwitchEnum;
  remark?: string;

  user?: IUser;

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

// ↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑ 直播间类型 ↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑

// ↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓ 用户类型 ↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓

export interface IQqUser {
  id?: number;
  client_id?: number;
  openid?: string;
  unionid?: string;
  nickname?: string;
  figureurl?: string;
  figureurl_1?: string;
  figureurl_2?: string;
  figureurl_qq_1?: string;
  figureurl_qq_2?: string;
  constellation?: string;
  gender?: string;
  city?: string;
  province?: string;
  year?: string;
  ret?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IWechatUser {
  id?: number;
  appid?: string;
  openid?: string;
  nickname?: string;
  sex?: number;
  province?: string;
  city?: string;
  country?: string;
  headimgurl?: string;
  privilege?: string;
  unionid?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IEmailUser {
  id?: number;
  email?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: any;
}

export interface IGithubUser {
  id?: number;
  client_id?: string;
  login?: string;
  github_id?: number;
  node_id?: string;
  avatar_url?: string;
  gravatar_id?: string;
  url?: string;
  html_url?: string;
  type?: string;
  site_admin?: string;
  name?: string;
  company?: string;
  blog?: string;
  location?: string;
  email?: any;
  hireable?: any;
  bio?: string;
  twitter_username?: any;
  public_repos?: number;
  public_gists?: number;
  followers?: number;
  following?: number;
  github_created_at?: string;
  github_updated_at?: string;
  private_gists?: number;
  total_private_repos?: number;
  owned_private_repos?: number;
  disk_usage?: number;
  collaborators?: number;
  two_factor_authentication?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: any;
}

export interface IThirdUser {
  id?: number;
  user_id?: number;
  third_user_id?: number;
  third_platform?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

/** 用户状态 */
export enum UserStatusEnum {
  /** 正常 */
  normal,
  /** 禁用 */
  disable,
}

export enum QuizMatchStatus {
  notStarted = 0, // 未开始
  inProgress = 1, // 进行中
  completed = 2, // 已结束
}

export enum UserIsTouristEnum {
  yes,
  no,
}

export interface IUser {
  id?: number;
  username?: string;
  password?: string;
  status?: UserStatusEnum;
  avatar?: string;
  desc?: string;
  token?: string;
  agent_account_for?: number;
  is_tourist?: UserIsTouristEnum;

  qq_users?: IQqUser[];
  wechat_users?: IWechatUser[];
  github_users?: IGithubUser[];
  email_users?: IEmailUser[];

  roles?: any[];
  auths?: any[];
  is_admin?: SwitchEnum;
  is_agent?: SwitchEnum;

  user_roles?: number[];
  wallet?: IWallet;

  live_room?: ILiveRoom;
  live_rooms?: ILiveRoom[];

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export enum FormTypeEnum {
  'input' = 'input',
  'password' = 'password',
  'number' = 'number',
  'select' = 'select',
  'radio' = 'radio',
  'checkbox' = 'checkbox',
  'markdown' = 'markdown',
  'switch' = 'switch',
  'upload' = 'upload',
  'treeSelect' = 'treeSelect',
  'datePickerRange' = 'datePickerRange',
  'datePickerDatetime' = 'datePickerDatetime',
}
// ...
export interface IWallet {
  id?: number;
  user_id?: number;
  balance?: string;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export enum ClientEnvEnum {
  android = 'android',
  ios = 'ios',
  ipad = 'ipad',
  web = 'web',
  web_mobile = 'web_mobile',
  web_pc = 'web_pc',
  windows = 'windows',
  macos = 'macos',
  linux = 'linux',
}

export enum ClientAppEnum {
  billd_live = 'billd_live',
  billd_live_admin = 'billd_live_admin',
  billd_desk = 'billd_desk',
  billd_desk_admin = 'billd_desk_admin',
}
// ...
export enum SwitchEnum {
  yes,
  no,
}